move auth login into separate service
