+++ ng-sandbox/src/app/components/auth/auth.component.ts
@@ -2,11 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { LoginType } from 'src/app/models/loginType';
 import { User } from 'src/app/models/user';
-import { DialogService } from 'src/app/services/dialog.service';
-import { ToastMessageService } from 'src/app/services/toast-message.service';
+import { AuthService } from 'src/app/services/auth.service';
 import { AppState } from 'src/app/store/app.reducer';
-import { LogOut, SetUser } from 'src/app/store/auth.actions';
-import { SetGames } from 'src/app/store/boardgames.actions';
 
 @Component({
   selector: 'app-auth',
@@ -15,42 +12,26 @@ import { SetGames } from 'src/app/store/boardgames.actions';
 })
 export class AuthComponent implements OnInit {
   constructor(
-    private toastSrv: ToastMessageService,
-    public dialogServ: DialogService,
-    private store: Store<AppState>
+    private store: Store<AppState>,
+    private authService: AuthService,
   ) {}
 
   loggedInUser: User | null = null;
   loginType = LoginType;
 
   ngOnInit(): void {
-    this.initUser();
+    this.authService.initUser();
     this.store.subscribe((state) => {
       this.loggedInUser = state.auth.user;
     });
   }
 
   login(type: LoginType) {
-    this.dialogServ.openLogin(type);
+    this.authService.login(type);
   }
 
   logOut() {
-    this.store.dispatch(new LogOut());
-    localStorage.removeItem('user');
-    this.store.dispatch(new SetGames([]));
-    this.toastSrv.showSuccessMessage('User is logged out');
+    this.authService.logOut();
   }
 
-  initUser() {
-    const data = localStorage.getItem('user');
-    if (data) {
-      const userData = JSON.parse(data);
-      const user = new User(userData.email, userData.id, userData.createdDateStamp);
-      if (user.isUserExpired()) {
-        localStorage.removeItem('user');
-      } else {
-        this.store.dispatch(new SetUser(user));
-      }
-    }
-  }
 }
